exercise-08/requirements.txt
google-cloud-bigquery==3.11.4
pandas==2.1.4
db-dtypes==1.1.1
matplotlib==3.8.2
jupyter==1.0.0
ipykernel==6.28.0